forcing symlinks always (needed at configure time)

+++ setup.py
@@ -168,32 +168,34 @@ def setup_package():
             setup(**setup_dict)
         else:
             # actual build, numpy.distutils-based
-            from numpy.distutils.command.build import build
-            class build_with_submodules(build):
-                def run(self):
-                    if os.path.exists('.git'):
-                        from numpy.distutils.exec_command import exec_command
-                        exec_command(['git', 'submodule', 'init'])
-                        exec_command(['git', 'submodule', 'update'])
-                        build.run(self)
-                        # now build symbolic links to repositories
-                        os.symlink(os.path.abspath('clawutil/src/python/clawutil'),
-                                                   'clawpack/clawutil')
-                        os.symlink(os.path.abspath('riemann/src/python/riemann'),
-                                                   'clawpack/riemann')
-                        # need this one to build Fortran sources naturally
-                        os.symlink(os.path.abspath('riemann/src'),
-                                                   'clawpack/riemann/src')
-                        os.symlink(os.path.abspath('visclaw/src/python/visclaw'),
-                                                   'clawpack/visclaw')
-                        os.symlink(os.path.abspath('pyclaw/src/pyclaw'),
-                                                   'clawpack/pyclaw')
-                        os.symlink(os.path.abspath('pyclaw/src/petclaw'),
-                                                   'clawpack/petclaw')
-                        
+            if os.path.exists('.git'):
+                from numpy.distutils.exec_command import exec_command
+                exec_command(['git', 'submodule', 'init'])
+                exec_command(['git', 'submodule', 'update'])
+                # *always* need these
+                # now build symbolic links to repositories
+                if not os.path.exists('clawpack/clawutil'):
+                    os.symlink(os.path.abspath('clawutil/src/python/clawutil'),
+                               'clawpack/clawutil')
+                if not os.path.exists('clawpack/riemann'):
+                    os.symlink(os.path.abspath('riemann/src/python/riemann'),
+                               'clawpack/riemann')
+                    # need this one to build Fortran sources naturally
+                if not os.path.exists('clawpack/riemann/src'):
+                    os.symlink(os.path.abspath('riemann/src'),
+                               'clawpack/riemann/src')
+                if not os.path.exists('clawpack/visclaw'):
+                    os.symlink(os.path.abspath('visclaw/src/python/visclaw'),
+                               'clawpack/visclaw')
+                if not os.path.exists('clawpack/pyclaw'):
+                    os.symlink(os.path.abspath('pyclaw/src/pyclaw'),
+                               'clawpack/pyclaw')
+                if not os.path.exists('clawpack/petclaw'):
+                    os.symlink(os.path.abspath('pyclaw/src/petclaw'),
+                               'clawpack/petclaw')
+
             from numpy.distutils.core import setup
-            setup(cmdclass={"build": build_with_submodules},
-                  configuration=configuration,
+            setup(configuration=configuration,
                   **setup_dict)
     except Exception as err:
         print err
